Added database for user information
Added Sqlite to all the projects to create a database for all of the user sign in information. Information that is entered on the registration page is added to the database assigned to variables in the form of a table. Should allow for an easy file to search for user information when they login.